Fix Copy(From|To)InterleavedBuffer() to not use channel numbers.
Instead, use the order of the provided ASIOBufferInfo collection.
Otherwise it's impossible to use these functions if the channel
layout of the interleaved buffer is not the same as the one
described by the contents of ASIOBufferInfo.

This fixes a crash when using record/playback files along with
input/output channels that are not consecutive from zero.